Bug 1673553 part 47 - Move AutoSuppressAllocationMetadataBuilder from TypeInference-inl.h to JSObject-inl.h. r=iain
The TypeInference-inl.h include in ObjectOperations-inl.h had to be removed to
break an include cycle. Unfortunately that was a load bearing include (the
TypeInference files pull in a lot of headers) so this adds various missing includes.

Depends on D98150

Differential Revision: https://phabricator.services.mozilla.com/D98151
